+++ scaffold-std-demo-app/src/test/java/space/x9x/demo/test/infrastructure/dao/RaffleActivityMapperTest.java
@@ -1,12 +1,14 @@
 package space.x9x.demo.test.infrastructure.dao;
 
+import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import space.x9x.demo.domain.manage.model.entity.RaffleActivityPageQuery;
 import space.x9x.demo.infrastructure.dao.mysql.IRaffleActivityMapper;
+import space.x9x.demo.infrastructure.dao.po.RaffleActivityPO;
 import space.x9x.radp.commons.json.JacksonUtils;
+import space.x9x.radp.spring.framework.dto.PageParam;
 
 /**
  * @author x9x
@@ -26,9 +28,12 @@ class RaffleActivityMapperTest {
 
     @Test
     void test_page() {
-        RaffleActivityPageQuery pageQuery = new RaffleActivityPageQuery();
-        pageQuery.setPageIndex(-1);
-        pageQuery.setPageSize(3);
-        log.info(JacksonUtils.toJSONString(raffleActivityMapper.pages(pageQuery, null)));
+        log.info(JacksonUtils.toJSONString(raffleActivityMapper.pages(new PageParam(1, 5), null)));
+    }
+
+    @Test
+    void test_pageXml() {
+        Page<RaffleActivityPO> filterPage = new Page<>(1, 5);
+        log.info(JacksonUtils.toJSONString(raffleActivityMapper.pagesXml(filterPage)));
     }
 }